requirements.txt
streamlit>=1.28.0
supabase>=2.0.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
cryptography>=41.0.0
python-dateutil>=2.8.2